Assignment_8/requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
oauth2client==4.1.3
pandas==2.1.3
aiohttp==3.9.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sse-starlette==1.6.5
nest-asyncio==1.5.8 